+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-# Version 1.0.0
-- Livrable du stage
+# Version 1.2.0
+- Modification de la rubrique modifier élève HTML, ajout du champ Niveau.
+- Modification de la rubrique enregistrer élève JS, Update Niveau et mise a null des infos de scolarité + delete participation classe verte.
 
 # Version 1.1.0
 - Ajout de la rubrique commission, permettant les checks sur les enfants.- Ajout de la rubrique confirmation, permettant les checks sur les enfants.
 - Modification de la restriction obligeant l'information enseignant dans l'enregistrement de classe.
 - Modification dans la proposition des niveau supérieur de classe
+
+# Version 1.0.0
+- Livrable du stage